Add confirmation for opening time/place discussion
