+++ apps/eventer_web/assets/js/src/features/eventPage/DecisionResolveForm.tsx
@@ -2,7 +2,7 @@ import { Form, Formik } from 'formik';
 import React from 'react';
 import DatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
-import { objectiveT } from './util';
+import { objectiveT } from './stateTransformations';
 
 type resolutionT = Date | string;
