+++ apps/eventer_web/assets/js/src/features/eventPage/useChannel.ts
@@ -12,7 +12,7 @@ import {
   stateEventT,
   addStateDecision,
   removeStateDecision,
-} from './util';
+} from './stateTransformations';
 import { useParams } from 'react-router-dom';
 
 type useChannelT = (
