+++ apps/eventer_web/assets/js/src/pages/EventPage.tsx
@@ -7,10 +7,12 @@ import BasicEventData from '../features/eventPage/BasicEventData';
 import Decision from '../features/eventPage/Decision';
 import Decisions from '../features/eventPage/Decisions';
 import EventUpdateForm from '../features/eventPage/EventUpdateForm';
+import OpenDiscussionConfirmation from '../features/eventPage/OpenDiscussionConfirmation';
 import RemoveDecisionConfirmation from '../features/eventPage/RemoveDecisionConfirmation';
+import { stateEventT } from '../features/eventPage/stateTransformations';
 import useChannel from '../features/eventPage/useChannel';
 import useChannelCallbacks from '../features/eventPage/useChannelCallbacks';
-import { stateEventT } from '../features/eventPage/util';
+import { hasExistingDecision } from '../features/eventPage/util';
 
 type eventUpdateFormModalChildT = {
   component: 'EventUpdateForm';
@@ -30,11 +32,17 @@ type removeDecisionModalChildT = {
   id: number;
 };
 
+type openDiscussionModalChildT = {
+  component: 'OpenDiscussionConfirmation';
+  objective: 'time' | 'place';
+};
+
 type modalChildT =
   | eventUpdateFormModalChildT
   | addDecisionModalChildT
   | decisionModalChildT
-  | removeDecisionModalChildT;
+  | removeDecisionModalChildT
+  | openDiscussionModalChildT;
 
 const EventPage: React.FC = () => {
   const { token, id: currentUserId } = useAuthorizedUser();
@@ -82,6 +90,16 @@ const EventPage: React.FC = () => {
     [setModalChild, setShouldShowModal],
   );
 
+  const showOpenDiscussionModal = useCallback(
+    (objective: 'time' | 'place') => {
+      setModalChild(
+        Some({ component: 'OpenDiscussionConfirmation', objective }),
+      );
+      setShouldShowModal(true);
+    },
+    [setModalChild, setShouldShowModal],
+  );
+
   const hideModal = useCallback(() => {
     setModalChild(None);
     setShouldShowModal(false);
@@ -101,8 +119,8 @@ const EventPage: React.FC = () => {
                 event={event}
                 currentUserId={currentUserId}
                 onEditEventClick={showEditModal}
-                onDiscussTimeClick={openTimeDiscussion}
-                onDiscussPlaceClick={openPlaceDiscussion}
+                onDiscussTimeClick={() => showOpenDiscussionModal('time')}
+                onDiscussPlaceClick={() => showOpenDiscussionModal('place')}
                 joinEvent={joinEvent}
                 leaveEvent={leaveEvent}
               />
@@ -147,6 +165,23 @@ const EventPage: React.FC = () => {
                             id={child.id}
                             title={decisions[child.id].title}
                             onConfirm={removeDecision}
+                            closeModal={hideModal}
+                          />
+                        );
+                      case 'OpenDiscussionConfirmation':
+                        const decisionExists = hasExistingDecision(
+                          decisions,
+                          child.objective,
+                        );
+                        return (
+                          <OpenDiscussionConfirmation
+                            objective={child.objective}
+                            hasCorrespondingDecision={decisionExists}
+                            onConfirm={
+                              child.objective === 'time'
+                                ? openTimeDiscussion
+                                : openPlaceDiscussion
+                            }
                             closeModal={hideModal}
                           />
                         );
